renaming the lambda files
