+++ efcms-service/src/cases/getCase.lambda.js
