+++ efcms-service/src/cases/getCase.lambda.test.js
@@ -1,6 +1,6 @@
 const expect = require('chai').expect;
 const lambdaTester = require('lambda-tester');
-const getCase = require('./getCase');
+const getCase = require('./getCase.lambda');
 const client = require('ef-cms-shared/src/persistence/dynamodbClientService');
 const sinon = require('sinon');
 const chai = require('chai');
