+++ efcms-service/src/documents/downloadPolicyUrl.lambda.js
