+++ efcms-service/src/documents/getUploadPolicy.lambda.js
