+++ efcms-service/src/documents/getUploadPolicy.lambda.test.js
@@ -19,7 +19,7 @@ const MOCK_RESPONSE = {
 };
 
 const getUploadPolicyStub = sinon.stub();
-const getUploadPolicy = proxyquire('./getUploadPolicy', {
+const getUploadPolicy = proxyquire('./getUploadPolicy.lambda', {
   '../applicationContext': {
     getPersistenceGateway: () => ({
       getUploadPolicy: getUploadPolicyStub,
