+++ efcms-service/src/swagger/swagger.lambda.js
